Add special JS variables that can be accessed from plugin configs
This allows postcss_binary invocations to refer to data files and
additional outputs using the correct paths without needing to change
directories or manually write workspace paths. I considered using
Blaze's make variable substitution for this instead, but that doesn't
work for outputs: the substitution is done in postcss_gen_runner, at
which point Bazel doesn't know which files are outputs and so doesn't
know to return their output location.